Sensors/requirements.txt
Flask
psycopg2
python-dotenv
requests
pymongo
